# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Matibabu EMR Lite API
  description: |
    REST API for the Matibabu Offline-First EMR Lite system.
    Provides endpoints for patient registration, retrieval, search, listing, updates, and deletion.
  version: 1.0.0
  contact:
    name: Matibabu Development Team
servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Patients
    description: Patient registry operations

paths:
  /api/patients:
    post:
      tags:
        - Patients
      summary: Register a new patient
      description: Registers a new patient record with demographic information. Generates a UUIDv7 identifier and timestamps.
      operationId: registerPatient
      requestBody:
        description: Patient demographic information for registration
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPatientRequest'
            example:
              firstName: John
              lastName: Kamau
              dateOfBirth: "1995-06-15"
              gender: MALE
              phoneNumber: "+254712345678"
              address: Nairobi
      responses:
        '201':
          description: Patient successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
              example:
                id: 018d3f1a-6379-7a0e-971c-fcae741cfb12
                firstName: John
                lastName: Kamau
                dateOfBirth: "1995-06-15"
                gender: MALE
                phoneNumber: "+254712345678"
                address: Nairobi
                createdAt: "2026-08-20T12:00:00Z"
                updatedAt: "2026-08-20T12:00:00Z"
        '400':
          description: Invalid request payload or validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '409':
          description: A patient with the specified phone number already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Patient with phone number already exists: +254712345678"

    get:
      tags:
        - Patients
      summary: List patients
      description: Retrieves a paginated list of registered patients.
      operationId: listPatients
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          example: 0
        - name: size
          in: query
          description: Number of records per page
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
          example: 20
        - name: sort
          in: query
          description: Sorting criteria in the format `property,(asc|desc)`. Default sort direction is ascending.
          required: false
          schema:
            type: array
            items:
              type: string
          example: ["createdAt,desc"]
      responses:
        '200':
          description: Paginated collection of patients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagePatientResponse'
              example:
                content:
                  - id: 018d3f1a-6379-7a0e-971c-fcae741cfb12
                    firstName: John
                    lastName: Kamau
                    dateOfBirth: "1995-06-15"
                    gender: MALE
                    phoneNumber: "+254712345678"
                    address: Nairobi
                    createdAt: "2026-08-20T12:00:00Z"
                    updatedAt: "2026-08-20T12:00:00Z"
                pageable:
                  pageNumber: 0
                  pageSize: 20
                  sort:
                    empty: true
                    sorted: false
                    unsorted: true
                  offset: 0
                  paged: true
                  unpaged: false
                totalElements: 1
                totalPages: 1
                last: true
                size: 20
                number: 0
                sort:
                  empty: true
                  sorted: false
                  unsorted: true
                numberOfElements: 1
                first: true
                empty: false

  /api/patients/search:
    get:
      tags:
        - Patients
      summary: Search patient by phone number
      description: Searches and retrieves a single patient record by their unique phone number.
      operationId: searchPatientByPhoneNumber
      parameters:
        - name: phoneNumber
          in: query
          description: Unique contact phone number to search for
          required: true
          schema:
            type: string
            pattern: '^\+?[0-9]{7,15}$'
          example: "+254712345678"
      responses:
        '200':
          description: Patient retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
              example:
                id: 018d3f1a-6379-7a0e-971c-fcae741cfb12
                firstName: John
                lastName: Kamau
                dateOfBirth: "1995-06-15"
                gender: MALE
                phoneNumber: "+254712345678"
                address: Nairobi
                createdAt: "2026-08-20T12:00:00Z"
                updatedAt: "2026-08-20T12:00:00Z"
        '400':
          description: Missing or invalid phone number query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Parameter 'phoneNumber' is required"
        '404':
          description: No patient found with the specified phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Patient not found with phone number: +254712345678"

  /api/patients/{id}:
    get:
      tags:
        - Patients
      summary: Get patient by ID
      description: Retrieves demographic details of a single patient by their unique identifier (UUIDv7).
      operationId: getPatientById
      parameters:
        - name: id
          in: path
          description: Unique patient identifier (UUID)
          required: true
          schema:
            type: string
            format: uuid
          example: 018d3f1a-6379-7a0e-971c-fcae741cfb12
      responses:
        '200':
          description: Patient details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
              example:
                id: 018d3f1a-6379-7a0e-971c-fcae741cfb12
                firstName: John
                lastName: Kamau
                dateOfBirth: "1995-06-15"
                gender: MALE
                phoneNumber: "+254712345678"
                address: Nairobi
                createdAt: "2026-08-20T12:00:00Z"
                updatedAt: "2026-08-20T12:00:00Z"
        '404':
          description: Patient not found with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Patient not found with id: 018d3f1a-6379-7a0e-971c-fcae741cfb12"

    put:
      tags:
        - Patients
      summary: Update patient demographic information
      description: Updates demographic information of an existing patient. Updates the `updatedAt` timestamp upon successful modification.
      operationId: updatePatient
      parameters:
        - name: id
          in: path
          description: Unique patient identifier (UUID)
          required: true
          schema:
            type: string
            format: uuid
          example: 018d3f1a-6379-7a0e-971c-fcae741cfb12
      requestBody:
        description: Updated patient demographic information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequest'
            example:
              firstName: John
              lastName: Kamau
              dateOfBirth: "1995-06-15"
              gender: MALE
              phoneNumber: "+254700000000"
              address: Mombasa
      responses:
        '200':
          description: Patient successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
              example:
                id: 018d3f1a-6379-7a0e-971c-fcae741cfb12
                firstName: John
                lastName: Kamau
                dateOfBirth: "1995-06-15"
                gender: MALE
                phoneNumber: "+254700000000"
                address: Mombasa
                createdAt: "2026-08-20T12:00:00Z"
                updatedAt: "2026-08-20T12:05:00Z"
        '400':
          description: Invalid request payload or validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Patient not found with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Patient not found with id: 018d3f1a-6379-7a0e-971c-fcae741cfb12"
        '409':
          description: Phone number is already registered to another patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Patient with phone number already exists: +254700000000"

    delete:
      tags:
        - Patients
      summary: Delete a patient
      description: Deletes an existing patient record by ID.
      operationId: deletePatient
      parameters:
        - name: id
          in: path
          description: Unique patient identifier (UUID)
          required: true
          schema:
            type: string
            format: uuid
          example: 018d3f1a-6379-7a0e-971c-fcae741cfb12
      responses:
        '204':
          description: Patient successfully deleted
        '404':
          description: Patient not found with the specified ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Patient not found with id: 018d3f1a-6379-7a0e-971c-fcae741cfb12"

components:
  schemas:
    Gender:
      type: string
      description: Gender of the patient
      enum:
        - MALE
        - FEMALE
        - OTHER
        - UNKNOWN
      example: MALE

    RegisterPatientRequest:
      type: object
      description: Request payload for registering a new patient
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
        - address
      properties:
        firstName:
          type: string
          description: Patient's first name (must not be blank)
          example: John
        lastName:
          type: string
          description: Patient's last name (must not be blank)
          example: Kamau
        dateOfBirth:
          type: string
          format: date
          description: Patient's date of birth (must be in the past or present)
          example: "1995-06-15"
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
          pattern: '^\+?[0-9]{7,15}$'
          description: "Patient's contact phone number (must be unique and valid: 7-15 digits with optional '+' prefix)"
          example: "+254712345678"
        address:
          type: string
          description: Patient's residential address (must not be blank)
          example: Nairobi

    UpdatePatientRequest:
      type: object
      description: Request payload for updating patient demographic information
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
        - address
      properties:
        firstName:
          type: string
          description: Patient's updated first name (must not be blank)
          example: John
        lastName:
          type: string
          description: Patient's updated last name (must not be blank)
          example: Kamau
        dateOfBirth:
          type: string
          format: date
          description: Patient's updated date of birth (must be in the past or present)
          example: "1995-06-15"
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
          pattern: '^\+?[0-9]{7,15}$'
          description: "Patient's updated contact phone number (must be unique and valid: 7-15 digits with optional '+' prefix)"
          example: "+254700000000"
        address:
          type: string
          description: Patient's updated residential address (must not be blank)
          example: Mombasa

    PatientResponse:
      type: object
      description: Demographic and metadata response model for a patient
      required:
        - id
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
        - address
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique patient identifier (UUIDv7)
          example: 018d3f1a-6379-7a0e-971c-fcae741cfb12
        firstName:
          type: string
          description: Patient's first name
          example: John
        lastName:
          type: string
          description: Patient's last name
          example: Kamau
        dateOfBirth:
          type: string
          format: date
          description: Patient's date of birth
          example: "1995-06-15"
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
          description: Patient's contact phone number
          example: "+254712345678"
        address:
          type: string
          description: Patient's residential address
          example: Nairobi
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the patient record was created
          example: "2026-08-20T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the patient record was last updated
          example: "2026-08-20T12:00:00Z"

    PagePatientResponse:
      type: object
      description: Paginated response containing a list of patients
      required:
        - content
        - totalPages
        - totalElements
        - last
        - first
        - size
        - number
        - numberOfElements
        - empty
      properties:
        content:
          type: array
          description: List of patients on the current page
          items:
            $ref: '#/components/schemas/PatientResponse'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        totalPages:
          type: integer
          description: Total number of pages available
          example: 1
        totalElements:
          type: integer
          format: int64
          description: Total number of records across all pages
          example: 1
        last:
          type: boolean
          description: Whether this is the last page
          example: true
        first:
          type: boolean
          description: Whether this is the first page
          example: true
        size:
          type: integer
          description: Requested page size
          example: 20
        number:
          type: integer
          description: Current page number (zero-based)
          example: 0
        sort:
          $ref: '#/components/schemas/SortObject'
        numberOfElements:
          type: integer
          description: Number of elements in the current page
          example: 1
        empty:
          type: boolean
          description: Whether the page contains no records
          example: false

    PageableObject:
      type: object
      description: Pagination metadata
      properties:
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 20
        sort:
          $ref: '#/components/schemas/SortObject'
        offset:
          type: integer
          format: int64
          example: 0
        paged:
          type: boolean
          example: true
        unpaged:
          type: boolean
          example: false

    SortObject:
      type: object
      description: Sorting metadata
      properties:
        empty:
          type: boolean
          example: true
        sorted:
          type: boolean
          example: false
        unsorted:
          type: boolean
          example: true

    ErrorResponse:
      type: object
      description: Standard error response
      required:
        - error
      properties:
        error:
          type: string
          description: Description of the error
          example: "Patient not found with id: 018d3f1a-6379-7a0e-971c-fcae741cfb12"

    ValidationErrorResponse:
      type: object
      description: Standard validation error response
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-08-20T12:00:00Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed for object='registerPatientRequest'"
        path:
          type: string
          example: "/api/patients"